Allow loading and configuring Brown's TNN sensors
Adapts the existing Hohonu framework to play nicer with multiple sources of data, as long as they can be converted into NAVD88. 

In this case it adds a new data source for the EPSCOR project that has sensors deployed using The Things Network. This uses the storage service to allow us to query up to a month's worth of data to build the QARTOD and device configurations.